Hyper-V 2016 actually uses 'the object is in use'
https://progress.opensuse.org/issues/41144

`hyperv_cmd_with_retry` is used in situations where the Hyper-V command
fails because the image is still downloaded by another job but we
attempted to use it.

It was a mistake to expect 'used by another process' in such cases 'The
operation cannot be performed while the object is in use.' (the actual
default here) should be used instead. Now, `Start-VM` command will be
re-triggered after some time.