boot_from_pxe: Be even more forgiving with slow installation startup
As we have seen in https://progress.opensuse.org/issues/41693 as well as
https://progress.opensuse.org/issues/31375
we do have some machines which would boot up after long time but do not
manage to complete within 300s which is already pretty long but given
that the initial bootup of the installation system also includes some
potential hardware discovery it might take longer depending on specific
hardware setups. We can be more forgiving here as installation is
assumed to be conducted not that often as boots/reboots of the installed
system.

Related progress issue: https://progress.opensuse.org/issues/41693