Run check_system_is_updated on sle
There were multiple issues with this module. First of all, we switch to
root-console, which is not expected in case of SLE as next test expects
tty with X11. Secondly, we run test after consoletest_setup, so command
fails as pkcon returns non-zero code, which is expected.
I tried to use script_output, but even if we use `proceed_on_failure` it
doesn't return any output, as command failed.

See [poo#31954](https://progress.opensuse.org/issues/31954).