[CLOUD] Use runltp-ng
Use the runltp-ng script from metan hackweek project to run ltp
on public cloud instance. Currently I'm on my fork, to have
needed changes in place. There are still some open questions,
like timeout handling. But basic functionality fits the need for
this test module and the runltp-ng script will be improved continuously.

The reboot.sh script is used by runltp-ng to recover the system,
whenever it hang or the kernel is tainted.