wicked: generic mutex approach
Need:
In advanced wicked testsuite, each sut testmodule has a corresponding
ref testmodule. These testmodules must be aligned at runtime.
We use mutex() to do so.

Problem:
When we just place the mutex_create() at the end of the run() method.
we face the problem, if a assert happen, that the sut and ref get
out of sync, cause ref still waits for the mutex_create().

We could place the mutex in post_fail_hook() and post_run_hook(). But
we need to make sure that the mutex is only called once.

Implementation approach:
Override basetest::done() to have a single execution point
at the end of a test_module and do the mutex stuff.
Use wicked_need_sync flag in tests_flags() to control, if a
testmodule which derive from wickedbase:: needs the mutex.