Work bsc#1103199 and bsc#1112109 around
We have a bug, which interferes with usage of serial console. On qemu
backend it's enough to wait for welcome message on tty, which means that
service is running. On Xen HVM service gets restarted constantly, so the
only possible workaround is to stop it ASAP.

See [poo#42359](https://progress.opensuse.org/issues/42359).