installation_overview: Simplify and cleanup
Previously installation_overview was unnecessarily wasting time with
`wait_still_screen` calls which are not needed anymore after we introduced the
initial assert_screen which checks for a completely loaded screen. Also,
obsolete checks for old bugs were included in the code but not even executed
since we do not have the according needles anymore since a very long time. The
same is true for 'inst-scrollbar' and 'manual-intervention'. However, I
consider it possible that a need for 'manual-intervention' comes back which we
then can simply make use of by creating the according needle. In this case I
assume the time behaviour would be different anyway so we can reduce the
timeout to save another 10s.

Verification run: http://lord.arch/tests/1901

Related progress issue: https://progress.opensuse.org/issues/42875