Allow to identify screens where license appeared
The commit adds two functions: one for identifying cases when installer
has to show Product Selection screen before the License Agreement and
the second to identify cases when License has to be shown on Welcome
screen.

The functions are added, because different SLE versions show the License
Agreement on diferent pages.

Made in scope of [poo#42608](https://progress.opensuse.org/issues/42608)