Introduce new flow for the validation of the installation
In many test suites we use standard set of test modules to validate that
SUT is functional. However, it's not efficient to run the very same set
of test modules for each installation. For example, if the only
difference of the setup is ext4 instead of btrfs, we need to validate
that FS works fine and risks that firefox won't start are quite low.

See [poo#42854](https://progress.opensuse.org/issues/42848).