Add select_{,user}_serial_terminal() helpers
These helpers are trying to select best available console.
The choice is made by BACKEND and other variables.
select_serial_terminal() is going to replace select_virtio_console().

virtio console is the default, variable NO_VIRTIO_CONSOLE=1 is needed
to disable it. This is different from select_virtio_console(), which
by default disabled and uses VIRTIO_CONSOLE=1 for enabling it.