wicked: Barriers on REF server (no support server)
The idea is to install support stuff like DHCP-server and openVPN
on the REF server it self.
So we can also use the REF server to create the barriers.
The REF Server need to be the Parent of SUT.
The barries are named like 'test_'.$testname.'_ready'.

Read the existing tests from filesystem. Benefit is, when adding
a new tests, we must not touch main_common.pm.

Introduced WICKED_EXCLUDE variable. This can be used to exclude
specific tests from the job. The given regex is matched against
the basename of the test.
Examples:
'^t0[56]_'        => run all tests instead of t05_* and t06_*
'^(?!t01_basic)'  => run only the test t01_basic