Separate main.pm for kubic
In a couple of months it is going to be a major codebase jump in
CaaSP, by start using Kubic as a standard development base. This
separation frees Kubic to move along into this direction much
faster and get up to speed in terms this architectural change.

In addition to that, community members do not know what CaaSP
is or what is running internally, so their PRs would be reviewed
much faster since there will be no risk in breaking SUSE QA/QAM.

Summary:

* Create a separate `main.pm` for Kubic
* Add `README.md`
* Prepare the system after inst. (boot, login, network, repo, etc)
* Increase the timeout for `rcshell`
* Remove original code from CaaSP that refers to Kubic.