Skip grub2 needle if system has already booted
The `process_reboot` is unreliable due to the fact that grub2
splascreen does not have a countdown timer. As a result there
might be some case where openQA waits for the grub2, while the
system has already passed that step and it waiting at the login
prompt. This PR fixes this by skipping the test for grub2 in
case grub splashscreen was too fast to check for.