Support pmem devices in partitioning_firstdisk
When running installation/partitioning_firstdisk, the test must
deselect all devices with the exception of /dev/sda (or similarly
named first disk device) to configure as the device where the OS
partition and swap will be located. This commit and related needles
allow the test to deselect pmem devices (/dev/pmem0 and /dev/pmem1
for the time being) so only /dev/sda is kept selected.