Accept addon licenses in SLES15 if registering after install
When adding modules to SLES15 during installation, most addons do not
show a license agreement and instead the process goes directly to the
registration code screen. There is already code in place in
lib/registration to handle addons with licenses and an array
SLE15_ADDONS_WITHOUT_LICENSE listing the addons that do not require a
license agreement on SLES15; however, there are addons included in that
array that do not require a license agreement during installation, but
do require one if registering the system after installation with 'yast2 scc'.

If registering after installation, license agreements are shown when
selecting the HA and WE modules, so the current code fails to accept
the licenses as it is using the same method for handling addons license
agreements as the one used during installation which is not expecting a
license agreement screen for these two modules.

In order to have a workaround so the same method can be used when adding
modules during or after installation, this commit adds and sets the env
variable IN_PATCH_SLE before calling the yast_scc_registration method
from tests/update/patch_sle, and also updates the accept_addons_license
method (itself called by yast_scc_registration) depending on the value of
that variable, so if yast_scc_registration is called while patching a system,
the code is able to properly handle the license screens from the HA and
WE modules.