publiccloud: Use temporal credentials
The idea is to use only temporal credentials in openQA. The
credentials are managed by an external daemon see:
https://github.com/cfconrad/pcw
It offers a REST api to list, create and delete such credentials.

This PR implements that logic only for EC2 as it is meant to be
the proposal. If we agree, we will implement the other providers.

The URL to connect to REST can be passed via
_SECRET_PUBLIC_CLOUD_REST_URL variable.