Add cluster hostnames to /etc/hosts if DNS is failing
In a 2-node cluster migration from SLES+HA 11-SP4, host name
resolution is required earlier than in the same test from
SLES+HA 12 or 15, because some configuration tasks need to be
executed in the cluster right after migration due to the changes
in the HA stack. Most HA tests rely on Multi Machine mode in
openQA and perform name resolution via the DNS service provided
by the support server, but in certain scenarios this service is
not available. This commit adds a workaround adding the hostnames
to /etc/hosts so that the test can continue when DNS name
resolution is failing.