Always try esc in post fail hook
It appears that we exit before trying to escape during a shutdown
scenario (or any other for that matter). @okurz's original comment even says
pressing esc may be useful during shutdown, but we have never done it.

In particular if the system refuses to shutdown and sysrq-d is pressed, we
need to delay for a little bit anyway so that QEMU is not closed before the
stuck tasks can be written to the serial.

For instances where it doesn't make sense to press escape, I don't think it
matters if we press it anyway. In fact we could randomly mash the keyboard (like
a real user) after failure, it might be useful.