Improve stability of yast2_snapper on weak machine
yas2_snapper module failed on the systems with low performance because
the test tried to close "Filesystem Snapshots" window by just sending a
shortcut without actually verifying if the window is closed or not.

The commit adds several attempts to close the window and to check if it
is actually closed and only then proceed further.

Related ticket: [poo#43784](https://progress.opensuse.org/issues/43784)