Boot and install on baremetal using ipxe and autoyast
When booting and installing on bare metal machines, the boot process
over the network and then using the ipmi console to control the boot
process was error prone and subject to a lot of failures
This is an initial effort to make the entire installation process more
robust for deployments on bare metal.

When installing using autoyast and ipxe, we don't need to use the
typical load_boot_tests(), load_inst_tests() and load_reboot_tests(),
but instead load a different test module, ipxe_install.

When using ipxe, uses_qa_net_hardware should return false, even if this
is QA network hardware. This way, we can easily avoid loading additional
test steps that conflict with what ipxe_install does.

Assumptions and external requirements:

It is assumed, an external http server is available, that implements the
following two methods:
  1. Method: POST. url: /<SUT_IP>/script.ipxe
     store a IPXE boot script
  2. Method: GET. url: /script.ipxe
     get the IPXE bootscript for the SUT (using the remote address.
A simple demo implementation used for development of this is available
at https://github.com/frankenmichl/ipxe_http

In addition, the machines that need to be installed with this method
need to be assigned a special IPXE binary, compiled to chainload the
script.ipxe from this http server.

Signed-off-by: Michael Moese <mmoese@suse.de>