Add utils subroutine to handle interactive command
The commands need to be interacted with user's input were usually run
and asserted by the way eg. combining "wait_serial" together with a
while loop, then match the prompt message in it. We want to simplify
this procedure by wrapping it in "script_run_interactive" subroutine.

Related progress issue: poo#45785

Also adjust the format for utils.pm to satisfy perltidy