Collect logs on 2nd stage of autoyast installation
openQA failed to collect logs if autoyast installation failed on the
second stage.

It was happened, because in parent post_fail_hook (in y2logsstep), the
'root-console' was selected, but black screen was shown instead of it,
so it might be unavailable there.

The commit allows to collect logs from 'install-shell' console.

Related ticket: [poo#42710](https://progress.opensuse.org/issues/42710)