Add check whether Hostnames window is closed
The test continued execution even if the 'YaST2 - Hostnames' window was
not closed (e.g. due to error) after pressing 'Ok' to save changes in
hosts configuration.

The commit adds additional check if the window is closed or not in order
to make the fail more obvious.

Related ticket: [poo#46232](https://progress.opensuse.org/issues/46232)