Modify interactive command function for AppArmor
A more robust subroution for interactive command has been already
added to the util.pm. Modify AppArmor cases with this one to
replace legacy function which is for AppArmor cases only in
lib/apparmortest.pm. At the same time, change the method of running
aa-genprof in background with the "screen" command, to make the test
process more stable.

Fix poo#45803