VMware: Add serial terminal for GRUB
https://progress.opensuse.org/issues/47150
https://trello.com/c/ABvmLsIh/346-p1-vmware-cant-connect-via-vnc-properly-to-guest-after-restart

On VMware, VNC connection is dropped on guest restart. We are able to
connect afterwards, but if we connect when the guest is not ready yet,
we get black screen, which won't refresh.

The solution is to add GRUB on serial console and checking for GRUB's
presence there. Once we detect GRUB on serial console, we know VNC is
ready as well and we can connect to `sut` console.

So that VNC output is not stalled, we need to "look aside" around the
time of restart (`svirt` console seems to be a good place to look at as
it's always present on svirt backend).

Validation runs:
* `default_install`: http://nilgiri.suse.cz/tests/229
* `textmode`: http://nilgiri.suse.cz/tests/230