Add PoC for scheduling with YAML files
As a part of [poo44420](https://progress.opensuse.org/issues/44420) we
continue to look for a solution for main.pm mess. This approach
simplifies what @okurz did in #6329.
As of now it forbids any conditions which are more complex than ingle
variable match.
This solution won't work straight away for bootloaders, as we have 7 of
them and only combination of 3 of them can decide which one to use. I
believe that's a good indicator that we either should separate those
schedules e.g. by using variable in the file name or having separate
test suite. I strongly believe if we use DASD in the installation it's
not the same test suite as on 64bit qemu. So this approach will force us
to answer those questions.
When trying to combine multiple things (like I did for TW and SLE), we
also will be forced to come up with namings for the conditions, which
will help us to ask ourselves why we have this difference at all and if
it should be there.

Approach can be easily extended to use single file for multiple test
suites, separate file for conditions (to reuse them).

If this solution will get positive feedback, we can start moving some
scenarios to use new scheduling mechanism and start solving issues which
prevent us from migration (e.g. "default" test suite is a mess and has
minor differences in ~10 setups we have for it).