Select sda if not selected in select_first_disk
When selecting the first disk for installation in a system with
disk and non-disk block devices, there exists the possibility that no
device is selected if the installer is pre-selecting a non-disk block
device by default. This commit ensures that sda is selected if it was
not selected by default.