Allow partition editing using combinations of data
The commit adds ability to edit suggested partitioning by defining a
different set of data from the test module.

This allows a test to select the specified options, depending on the
data that it provides to the method, instead of making different set
of methods for all the required test combinations.