Workaround for poo#44771
This is a workaround for poo#44771 "Can't call method "exec" on an
undefined value". It seems that 'svirt' console is not initialized at
this point (perhaps due to a previous `reset_console`) and thus may not
be used.

Validated by three runs on VMware and 2 runs on Hyper-V 2012 R2, where
it always fails (for me and OSD anyway).