Use actual serial device for aarch64 KVM grub configuration
This commit should be merged together with
https://github.com/os-autoinst/os-autoinst/pull/1107

There is a defect found during openQA integration with aarch64 automation.
Firstly, not all aarch64 machines use ttyAMA0 as their serial console.
Some vendor meachines like xgene has only ttyS0 as their serial console.
Secondly, the SERIALDEV variable is overridden by some other functionalities
during automation run. So it is necessary to save its orignal value to
SERIALDEVORIG variable which is only for internal use. It should not be
referred to from outside of live test run, for example, assign a value to it
explicitly in test suite or machine settings.

Author:    Wei Chen <wchen@suse.com>
Date:      Fri Feb 15 14:15:59 2019 +0800