Do not switch consoles to disable ipv6 and ipv4
The test nm wpa enterprise tries to switch to root console at some point
so that it can disable ipv6 and ipv4 just to trick nm that we're online
switching makes the test fail from time to time, due to the wrong tty
being selected or passwords not matching somehow. Instead open a
terminal in the x11 console and do everything from there.