Ensure the dependency of the junit converter is installed
Depending on the booted image python, dependency for the junit
converter, is installed or not. Trying to schedule a testrun based on
QA_TESTSET instead of QA_TESTSUITE showed the problem of missing python
interpreter as well as missing dependencies new for SLE15SP1 because of
packages moved to another module. This commit ensures the python
dependencies are installed which is not even harmful if it is already
installed in a previous step.

Related progress issue: https://progress.opensuse.org/issues/44138