Schedule orphaned_packages_check in more scenarios
orphaned_packages_check should be part of general installation
validation and not only for upgrades or migrations.

It now runs in more scenarios (default, extra_tests_in_textmode) but
there is also one that it does not run in anymore:
sle-15-SP1-Installer-DVD-x86_64-Build175.4-online_sles15_pscc_basesys+srv_def_full_y@64bit.
It still runs in other migration scenarios, so this should not be an
issue.