opensusebasetest: Fix export_kde_logs failing to type initial character
Explicitly switching the console can still cause a delay causing the
first characters to be lost. Starting with `type_string` does not
validate the command execution so the actual upload would fail in a
dubious way so it is better to use `assert_script_run` in the first
place.

Related progress issue: https://progress.opensuse.org/issues/35589