fix the infiniband tests
There were some problems with the infiniband testsuite that are now
solved.
A third barrier is needed to synchronize the setup parts in order to
avoid failures whith copying ssh keys when one machine is not yet booted
in time.

In addition, the documentation was updated and some dependencies we need
for minimal installations were added.

Signed-off-by: Michael Moese <mmoese@suse.de>