Fix SLES4SAP XRDP test on ppc64le - version 2
Sometimes screensaver can appears on the XRDP server during the test
and thus can avoid login at the end of the test.

This has already been fixed for x86_64, but it seems that we have to
click on the mouse before on ppc64le (bug? Will be investigated after).

This commit *tries* to fix this a second time.