Boot with kernel/initrd from repository if needed
The commit removes soft-fail for boo#990374 and turns it into
record_info to just highlight entering kexec parameter to bootloader.

The parameter is entered for the test, that is using LATEST TW snapshot
NET installer (to be tested, not yet published) and installs the
'previous' (currently published TW snapshot).

kexec value is set with LINUXRC_KEXEC openQA variable. Possible values
are 0,1,2,3.

Please, see the documentation for linuxrc and kexec usage
[here](https://en.opensuse.org/SDB:Linuxrc)