Fix unlocking desktop on Plasma 5.16+
In Plasma 5.16 (git master currently), the login and lock screen
behaviour changed, there's now a shorter timeout for the idle screen.

Previously, openQA would see the idle screen, press esc and then wait
for a screen change and then a still screen. This does not work with the
new behaviour as it practically waited for the idle screen again:
https://openqa.opensuse.org/tests/888171

Fix it by moving the delay before the esc keypress.