Change calls to sleep() for WebDriverWait()
This commit replaces most calls to time.sleep() with find_element(),
which calls for WebDriverWait() instead. This helps the test to wait
for the pages to load instead of just waiting for a set amount of time.
Changes were only done where applicable, as there are still some calls
to sleep() that need remain in the code (basically when the element to
search for in the page is already there but obscured).

As an added bonus, with the change the python script runs almost twice
as fast as before, so the 15 minutes timeout in ha/hawk_gui.pm has
been reduced to 10 minutes.