Remove use of lib/selenium methods in ha/hawk_gui
lib/selenium uses Selenium::Remote::Driver which is not currently installed
in openqa.suse.de workers. Since the only method from lib/selenium in
use in the ha/hawk_gui module is enable_selenium_port, which in turns is
only a call to assert_script_run('systemctl stop ...'), this substitutes
that call instead of the call to the method.