Use mstflint instead of Mellanox MFT
Building MFT utility from Mellanox was subject to failure from time to
time, and required frequent adjustment of Variables to pull the latest
version.
However, the same functionality is also provided by the tool "mstflint"
that is already available in the distribution.
However, the tool is slightly different and does not come with the
kernel module included in MFT - thus we need to use PCI address of the
devices instead of some static path.
In addition, as we change this code, we should enable support to
configure SR-IOV to enable the creation of virtual functions to be
passed to virtual machines running on this hardware. In order to enable
this, set MLX_SRIOV=1 and MLX_NUM_VFS to the number of virtual functions
wanted, for example MLX_NUM_VFS=32.

Signed-off-by: Michael Moese <mmoese@suse.de>