Detect typos in boot parameters in bootloader.pm
Compare boot parameters that are expected to be entered with the actual
ones in the serial output in bootloader.pm module.

Does not fail a job but just produces a notification to openQA using
record_info.

Added to make the review of typos in bootloader parameters more simple.

Related ticket: [poo#48572](https://progress.opensuse.org/issues/48572)