Wait for boot cycle to start after node fencing
During HA cluster tests, one of the nodes fences the other causing
it to reboot. Until now this has been handled by loading the boot test
modules after ha/fencing to handle the booting process before continuing
the HA tests, but on occasions it is possible for the fencing action to
take longer to be applied in the node, causing a timeout failure in
boot/boot_to_desktop which has a timeout ranging from 80 to 140 seconds
to wait for the grub2 or the boot screen.

This commit forces ha/fencing to wait until the root-console screen is
gone from the fenced node before continuing into the boot tests, allowing
the boot tests to actually only wait for the different booting screens.