Replace get_node_to_join with HA_CLUSTER_INIT
The method get_node_to_join() is not usable in the master node
as it requires the variable HA_CLUSTER_JOIN which should not be
defined in master node tests. This instead will check with
get_var('HA_CLUSTER_INIT') which is the variable partnered to
HA_CLUSTER_JOIN for the master node.