Remove return value logic from hawk_test
This removes the return value logic from hawk_test.py and
instead makes the script returns the total number of tests
minus the total number of passed tests. This means the script
still returns 0 on success, and a value greater than 0 if any
of the tests fail, but the return value itself does not have
any other meaning.