Combine all bootloaders in one module
Currently there are several bootloader test modules (bootloader, bootloader_zkvm,
bootloader_uefi etc.) that are scheduled in main.pm.

The conditional logic is too complex for declarative scheduling that is
planned to be applied to all YaST test suites.

In order to simplify migrating to declarative scheduling, as a simple and
straightforward workaround it is decided to extract the conditional logic to one
'bootloader' test module and use it there.

Related ticket: https://progress.opensuse.org/issues/50177