wicked: Use offline upload of log files
Use serial_file::upload_file() for uploading log files. This avoid the
need of working network after test.

Allow upload files via serial on first console:
If the test was run with VIRTIO_CONSOLE_NUM < 2, the upload should still
work.
The disadvantage is, that the file is visible on serial.txt in base64
format.