Support Server: add nfs server
In order to enable NFS server, the 'SUPPORT_SERVER_ROLES' should
contain 'nfs' entry. This will create default /nfs/share which can
be used on the client side, the test code.
The client code might want to change the dafault values, as the
testing might require - for instance - different permissions for
the shared directory. This can be done by setting test variable
CONFIGURE_NFS_SERVER to 1. This is meant to be further expanded
if needed, so various NFS versions might be used in different ways