Slurm: Test for slurm configured with db
Slurm can be configured with the database for 'accounting' and this
tests is meant to test this functionality of the slurm and the HPC
product. Slurm daemon for managing db access requires specific
slurmdb.conf. Also the naming (declarative names) for the HPC cluster
is changed to mater-slave convention, as this approach should allow
reusing the code easier.