zdup: Get rid of unchecked / not needed cdrkit install command
cdrkit-cdrtools-compat is either already installed on the systems we
care about or does not even exist. With `script_run` any error is
ignored so the call is pretty useless and should be removed when there
was no need seen in handling the install errors at all.

See https://openqa.suse.de/tests/2953674#step/zdup/6 for a migration job
from SLE12SP2 where the package is already installed. in contrast
https://openqa.opensuse.org/tests/951801#step/zdup/4 is an upgrade test
from openSUSE Leap 42.1 to openSUSE Tumbleweed ignoring the fact that
the test does not exist anymore. The same is true for SLE15.