Allow some settling time for runner to show the results
Currently, when the runner results does not show the results immediately
(default timeout is 0), the check fails, the runner is closed with 'esc',
and the system waits for 30 seconds.

Instead, give the runner a reasonable chance to present its results, and
only repeat the sequence if the results does not show up after the default
timeout of 30 seconds.

See poo#53045

Signed-off-by: Stefan Brüns <stefan.bruens@rwth-aachen.de>