KOTD: Add KOTD_REPO variable, change INSTALL_KOTD and openSUSE support
INSTALL_KOTD is now a flag which enables KOTD (otherwise it'd be normal
install_ltp test).

KOTD_REPO has similar function that previously KOTD_REPO had. But it was
simplified a bit: it's just URL or default value. Previous version was
URL or just a version. Version is not good as we keep to update it when
new version is released and version shifts which leads to testing old
repository. Useful is to have meaningful default and allow to specify full URL.
So now it's good enough just to set INSTALL_KOTD=1 and not set KOTD_REPO at all.

Adding openSUSE supports allows us to enable KOTD on o3.