Check timeout also when wiggling mouse on live CD
The timeout check was never done when the screenlock was detected once
and none of the needles match.

Also, the timeout would only start after the first screenlock, so replace
the assert_screen with a check_screen loop and reduce the remaining time
each time the check_screen has not matched.

Related issue: #53534