clean autoyast profile for mlx_con5
There is way too much information in this autoyast profile (it was generated
from Yast). We don't want to have user/group or even CUPS configuration, we
just need to ensure we use /dev/sda for installing, configure a bootloader
etc.
So let's remove all the unneeded information and improve maintainability
here, in most cases the defaults should work fine.