Remove existing ssh key before creating a new one
Sometimes, there are already existing ssh keys on the machine (either
from an old installation or from a previous test) and the creation of
the keys for this test does fail.
However, we do want to create new keys anyway and we don't want to
depend on the (non-)existance of ~/.ssh/id_rsa[.pub], we should just
remove the file if it is present.

Signed-off-by: Michael Moese <mmoese@suse.de>