hawk_test: Add click by xpath instead of partial link
Older versions of geckodriver and python are not very reliable
when clicking on the main page of hawk by partial link, specifically
on the Add Resources and Dashboard links, and in the Nodes tab.
This commit adds a click by xpath for those links, using the text
of the href code on the links.