HPC: Comprehensive HPC cluster tests
This test is aiming at comprehensive HPC testing. That means
that most components of the HPC product should be installed and
configured. At the same time it is expected that the HPC cluster
set-up is fairly complex, so that NFS shares are mounted, required
database(s) are ready to be used, some data is being populated to
the databases, artificial users are added etc.