Fix remove_grub_cmdline_settings() joining two strings
This fixes issue when removed string is in the middle of the text:

GRUB_CMDLINE_LINUX_DEFAULT="Hello foo world"

remove_grub_cmdline_settings("foo");
GRUB_CMDLINE_LINUX_DEFAULT="Helloworld"

This adds extra space, but that's not a problem for grub configuration.
We could be precise and not adding this space, but code simplicity wins.

NOTE: if we *want* to be precise we could:
1) Use look-ahead and look-behind, but that'd require change_grub_config
switch to use perl (sed does not support them).
2) Or keep using sed, but having possibility to pass $old and $new in
change_grub_config() also as arrays.

Both of them would allow to use more powerful regexes.

Reported-by: James Wang <jnwang@suse.com>