Check if we need to change configuration
Most of the time, we run our test against the same configuration for the
Mellanox boards. We should check first to see if we need to change the
configuration, and only do so if we have to.
The configuration itself takes a while, but requires a reboot afterwards,
so this can easily sum up to several minutes of time we can save.

Signed-off-by: Michael Moese <mmoese@suse.de>