Add ability to find all the nodes in xml by xpath
The commit adds the function to have ability to find all the nodes in
xml file by xpath and returns them as an array.

That allows to manipulate with the found nodes in the manner that is
required for the certain test (e.g. get values from the nodes, count
them etc.)