Fix regression introduced in #7802
In #7802 we have removed `assert_screen` call after we open firewall
details dialog. After that we open logs dialog, but press key to early
as previous dialog is still open.
With this PR, we assert that we got to the original screen before the
key press.

See [poo#53972](https://progress.opensuse.org/issues/53972).