LTP: Install runtime dependencies only for git installation
These dependencies were moved to packages as well, so it'll be faster
not install them also via perl.

Although not needed, just for sure added --recommends parameter to
zypper in command.

+ print more info about package (with rpm -qi pkg instead of rpm -q pkg)

NOTE: new dependencies needs to be added to install_ltp.pm (for git
installation) and to all 4 packages):
https://build.suse.de/package/view_file/QA:Head/qa_test_ltp/qa_test_ltp.spec?expand=1
https://build.suse.de/package/view_file/QA:Head/ltp/ltp.spec?expand=1
https://build.opensuse.org/package/view_file/benchmark/ltp/ltp.spec?expand=1
https://build.opensuse.org/package/view_file/benchmark:ltp:devel/ltp/ltp.spec?expand=1

Other option would be to try first install everything in single call and
if it fails run them one by one (suggested by Martin Doucha). But
disadvantage of this would be slowdown due slow typing on IPMI
(not only zypper itself is slow, but typing can be slow on slow consoles
- IPMI or non-virtio/serial on qemu/svirt).

See poo#54506, poo#56030