LTP: Fix tag name for verification builds via github PR
Github pull requests done via openqa-clone-custom-git-refspec script set
BUILD variable as user/branch-name#GITHUB_PR. BUILD variable is used as
part of tag (via get_ltp_tag()) which is used as filename. Thus we need
to replace special characters so that filename is valid. It'd be enough
to replace # and /, but just for case replace all non-alpha numeric
and _@ characters with - (dash).