Run fetches from registry.opensuse.org through a local proxy
Fetches from the registries have to be rewritten just like package
downloads from distro repos. While this approach is more complex and
more of a hack than the zypper repo replacement, it's a self-contained
process which works for all container engines and use cases.

If the REGISTRY_PREFIX variable is set, the provided regproxy service is
started and traffic to registry.opensuse.org tunneled through it.
(See regproxy/README.md for details).

The .pem files are part of this repo to allow for easier use with
multimachine testing in the future.