Allow import of several test_data in yaml schedule
The commit adds ability to use multiple !include tags in yaml scheduling
files as well as mixing !include tag with the test data that is defined
in the scheduling file directly. In case of the latter, the keys from
scheduling file have priority over imported ones with !include tag (i.e.
it is allowed to override the values defined in test data imported file)