Comment PR on staging test change
The commit adds the script that is executed on Travis run while PR is
created. The script checks if there are changes in test modules that are
used in staging test scenarios and puts a hint as a comment to the PR to
make a VR, confirming that the changes did not broke staging tests.