Fix SLES4SAP XRDP test
Connection on SLES4SAP host through XRDP from a Windows machine can
sometimes "fails" with a "RDP Services session has ended" error like on
the Windows VM.
In that case, connection through XRDP is OK but the failure is only
during the closing and only on the Windows side.

So this commit fix this by "simply" ack the "error" and continue.