publiccloud: Add single quote to run_ssh_command()
The normal intention by this function is, that the given command is
executed on the remote side. So if I create a redirection, the file
will appear on the remote side. To do so, we need to avoid bash special
characters to be interpreded. This is now done by single quote the
given command.